pyproject.toml
[project]
name = "football_pool"
version = "1.1.0"
description = "A Flask app for a website tracking the results/history of my family's football pool."
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.4",
    "flask-apscheduler>=1.13.1",
    "flask-migrate>=4.1.0",
    "psycopg[c]>=3.2.9",
    "pydantic>=2.11.7",
    "pytz>=2025.2",
    "requests>=2.32.4",
    "tzdata>=2025.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]

[tool.ruff]
line-length = 100
target-version = "py313"
include = ["apps", "migrations", "tests"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # Pyflakes
    "I",  # isort
    "W",  # pycodestyle warnings
    "C90", # mccabe complexity
]
ignore = []

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"] # unused imports

[tool.uv]
default-groups = "all"
link-mode = "copy"
